+++ myia/api.py
@@ -1,940 +1,29 @@
 """User-friendly interfaces to Myia machinery."""
 
 import inspect
-import math
-import numpy as np
-from types import FunctionType
-from collections import defaultdict
 
-from . import dtype, parser, composite as C, operations
-from .cconv import closure_convert
-from .dtype import Tuple, List, Class, Array, Int, Float, Bool, \
-    Number, tag_to_dataclass, ismyiatype, type_to_np_dtype, \
-    TypeMeta, Function
-from .infer import InferenceEngine, ANYTHING, Context, MyiaTypeError
-from .ir import Graph, clone, GraphManager
-from .opt import PatternEquilibriumOptimizer, lib as optlib, CSE, \
-    erase_class, erase_tuple
-from .pipeline import PipelineStep, PipelineResource, PipelineDefinition
-from .prim import py_implementations, vm_implementations, ops as P
+from .compile import step_wrap_primitives, step_compile, step_link, step_export
+from .infer import Context, MyiaTypeError
+from .ir import GraphManager
+from .pipeline import PipelineDefinition
+from .prim import py_implementations
 from .prim.value_inferrers import ValueTrack, value_inferrer_constructors
 from .prim.type_inferrers import TypeTrack, type_inferrer_constructors
 from .prim.shape_inferrers import ShapeTrack, shape_inferrer_constructors
-from .specialize import TypeSpecializer
-from .utils import TypeMap, as_frozen, overload, flatten, UNKNOWN
-from .vm import VM
-from .compile import step_wrap_primitives, step_compile, step_link, step_export
-from .validate import validate, whitelist as default_whitelist, \
-    validate_type as default_validate_type
-
-
-scalar_object_map = {
-    operations.add: P.scalar_add,
-    operations.sub: P.scalar_sub,
-    operations.mul: P.scalar_mul,
-    operations.truediv: C.scalar_truediv,
-    operations.floordiv: C.scalar_floordiv,
-    operations.mod: P.scalar_mod,
-    operations.pow: P.scalar_pow,
-    operations.eq: P.scalar_eq,
-    operations.ne: P.scalar_ne,
-    operations.lt: P.scalar_lt,
-    operations.gt: P.scalar_gt,
-    operations.le: P.scalar_le,
-    operations.ge: P.scalar_ge,
-    operations.pos: P.scalar_uadd,
-    operations.neg: P.scalar_usub,
-    operations.exp: P.scalar_exp,
-    operations.log: P.scalar_log,
-    operations.sin: P.scalar_sin,
-    operations.cos: P.scalar_cos,
-    operations.tan: P.scalar_tan,
-    operations.not_: P.bool_not,
-    operations.and_: P.bool_and,
-    operations.or_: P.bool_or,
-    operations.matmul: P.dot,
-    operations.getitem: C.getitem,
-    operations.setitem: C.setitem,
-    operations.bool: P.identity,
-    operations.getattr: P.getattr,
-    operations.setattr: P.setattr,
-    operations.len: C._len,
-    operations.make_tuple: P.make_tuple,
-    operations.make_list: P.make_list,
-    operations.iter: C.iter,
-    operations.hasnext: C.hasnext,
-    operations.next: C.next,
-    operations.to_array: C.to_array,
-    operations.switch: P.switch,
-    math.floor: P.scalar_floor,
-    math.trunc: P.scalar_trunc,
-    math.exp: P.scalar_exp,
-    math.log: P.scalar_log,
-    math.sin: P.scalar_sin,
-    math.cos: P.scalar_cos,
-    math.tan: P.scalar_tan,
-}
-
-
-standard_object_map = {
-    operations.add: C.add,
-    operations.sub: C.sub,
-    operations.mul: C.mul,
-    operations.truediv: C.truediv,
-    operations.floordiv: C.floordiv,
-    operations.mod: C.mod,
-    operations.pow: C.pow,
-    operations.eq: C.eq,
-    operations.ne: C.ne,
-    operations.lt: C.lt,
-    operations.gt: C.gt,
-    operations.le: C.le,
-    operations.ge: C.ge,
-    operations.pos: C.uadd,
-    operations.neg: C.usub,
-    operations.exp: C.exp,
-    operations.log: C.log,
-    operations.sin: C.sin,
-    operations.cos: C.cos,
-    operations.tan: C.tan,
-    operations.not_: C.not_,
-    operations.and_: C.and_,
-    operations.or_: C.or_,
-    operations.matmul: C.matmul,
-    operations.getitem: C.getitem,
-    operations.setitem: C.setitem,
-    operations.bool: C.bool,
-    operations.getattr: P.getattr,
-    operations.setattr: P.setattr,
-    operations.len: C._len,
-    operations.make_tuple: P.make_tuple,
-    operations.make_list: P.make_list,
-    operations.iter: C.iter,
-    operations.hasnext: C.hasnext,
-    operations.next: C.next,
-    operations.to_array: C.to_array,
-    operations.switch: P.switch,
-    math.floor: P.scalar_floor,
-    math.trunc: P.scalar_trunc,
-    math.exp: P.scalar_exp,
-    math.log: P.scalar_log,
-    math.sin: P.scalar_sin,
-    math.cos: P.scalar_cos,
-    math.tan: P.scalar_tan,
-    np.floor: C.floor,
-    np.trunc: C.trunc,
-    np.add: C.add,
-    np.subtract: C.sub,
-    np.multiply: C.mul,
-    np.divide: C.truediv,
-    np.mod: C.mod,
-    np.power: C.pow,
-    np.exp: C.exp,
-    np.log: C.log,
-    np.sin: C.sin,
-    np.cos: C.cos,
-    np.tan: C.tan,
-}
-
-
-standard_method_map = TypeMap({
-    dtype.Bool: {
-        '__and__': P.bool_and,
-        '__or__': P.bool_or,
-        '__eq__': P.bool_eq,
-        '__ne__': C.bool_ne,
-        '__bool__': P.identity,
-    },
-    dtype.Int: {
-        '__add__': P.scalar_add,
-        '__sub__': P.scalar_sub,
-        '__mul__': P.scalar_mul,
-        '__floordiv__': C.int_floordiv,
-        '__truediv__': C.int_truediv,
-        '__mod__': P.scalar_mod,
-        '__pow__': P.scalar_pow,
-        '__floor__': P.identity,
-        '__trunc__': P.identity,
-        '__pos__': P.scalar_uadd,
-        '__neg__': P.scalar_usub,
-        '__eq__': P.scalar_eq,
-        '__ne__': P.scalar_ne,
-        '__lt__': P.scalar_lt,
-        '__gt__': P.scalar_gt,
-        '__le__': P.scalar_le,
-        '__ge__': P.scalar_ge,
-        '__bool__': C.int_bool,
-        '__myia_to_array__': P.scalar_to_array,
-    },
-    dtype.UInt: {
-        '__add__': P.scalar_add,
-        '__sub__': P.scalar_sub,
-        '__mul__': P.scalar_mul,
-        '__floordiv__': P.scalar_div,
-        '__truediv__': C.int_truediv,
-        '__mod__': P.scalar_mod,
-        '__pow__': P.scalar_pow,
-        '__floor__': P.identity,
-        '__trunc__': P.identity,
-        '__pos__': P.scalar_uadd,
-        '__neg__': P.scalar_usub,
-        '__eq__': P.scalar_eq,
-        '__ne__': P.scalar_ne,
-        '__lt__': P.scalar_lt,
-        '__gt__': P.scalar_gt,
-        '__le__': P.scalar_le,
-        '__ge__': P.scalar_ge,
-        '__bool__': C.int_bool,
-        '__myia_to_array__': P.scalar_to_array,
-    },
-    dtype.Float: {
-        '__add__': P.scalar_add,
-        '__sub__': P.scalar_sub,
-        '__mul__': P.scalar_mul,
-        '__floordiv__': C.float_floordiv,
-        '__truediv__': P.scalar_div,
-        '__mod__': P.scalar_mod,
-        '__pow__': P.scalar_pow,
-        '__floor__': P.scalar_floor,
-        '__trunc__': P.scalar_trunc,
-        '__pos__': P.scalar_uadd,
-        '__neg__': P.scalar_usub,
-        '__eq__': P.scalar_eq,
-        '__ne__': P.scalar_ne,
-        '__lt__': P.scalar_lt,
-        '__gt__': P.scalar_gt,
-        '__le__': P.scalar_le,
-        '__ge__': P.scalar_ge,
-        '__bool__': C.float_bool,
-        '__myia_to_array__': P.scalar_to_array,
-    },
-    dtype.Tuple: {
-        '__len__': P.tuple_len,
-        '__getitem__': P.tuple_getitem,
-        '__setitem__': P.tuple_setitem,
-        '__myia_iter__': P.identity,
-        '__myia_next__': C.tuple_next,
-        '__myia_hasnext__': C.tuple_hasnext,
-    },
-    dtype.List: {
-        '__len__': P.list_len,
-        '__getitem__': P.list_getitem,
-        '__setitem__': P.list_setitem,
-        '__myia_iter__': C.list_iter,
-    },
-    dtype.Array: {
-        '__add__': C.array_add,
-        '__sub__': C.array_sub,
-        '__mul__': C.array_mul,
-        '__truediv__': C.array_truediv,
-        '__floordiv__': C.array_floordiv,
-        '__mod__': C.array_mod,
-        '__pow__': C.array_pow,
-        '__floor__': C.array_floor,
-        '__trunc__': C.array_trunc,
-        '__pos__': C.array_uadd,
-        '__neg__': C.array_usub,
-        '__eq__': C.array_eq,
-        '__ne__': C.array_ne,
-        '__lt__': C.array_lt,
-        '__gt__': C.array_gt,
-        '__le__': C.array_le,
-        '__ge__': C.array_ge,
-        '__matmul__': P.dot,
-        '__len__': P.array_len,
-        '__getitem__': P.array_getitem,
-        '__setitem__': P.array_setitem,
-        '__myia_iter__': C.array_iter,
-    },
-    dtype.JTagged: {
-    },
-    dtype.SymbolicKeyType: {
-    },
-    dtype.EnvType: {
-    },
-})
-
-
-@overload
-def default_convert(env, fn: FunctionType):
-    """Default converter for Python types."""
-    g = clone(parser.parse(fn))
-    env.object_map[fn] = g
-    return g
-
-
-@overload  # noqa: F811
-def default_convert(env, g: Graph):
-    mng = env.resources.manager
-    if g._manager is not mng:
-        g2 = clone(g)
-        env.object_map[g] = g2
-        return g2
-    else:
-        return g
-
-
-@overload  # noqa: F811
-def default_convert(env, seq: (tuple, list)):
-    return type(seq)(env(x) for x in seq)
-
-
-@overload  # noqa: F811
-def default_convert(env, x: (object, type, TypeMeta)):
-    return x
-
-
-############
-# Pipeline #
-############
-
-
-class _Unconverted:
-    # This is just used by Converter to delay conversion of graphs associated
-    # to operators or methods until they are actually needed.
-    def __init__(self, value):
-        self.value = value
-
-
-class Converter(PipelineResource):
-    """Convert a Python object into an object that can be in a Myia graph."""
-
-    def __init__(self, pipeline_init, object_map, converter):
-        """Initialize a Converter."""
-        super().__init__(pipeline_init)
-        self.converter = converter
-        self.object_map = {}
-        for k, v in object_map.items():
-            self.object_map[k] = _Unconverted(v)
-        for prim, impl in self.resources.py_implementations.items():
-            self.object_map[impl] = prim
-        type_map = {
-            bool: dtype.Bool,
-            int: dtype.Int,
-            float: dtype.Float,
-            np.ndarray: dtype.Array,
-            np.int8: dtype.Int,
-            np.int16: dtype.Int,
-            np.int32: dtype.Int,
-            np.int64: dtype.Int,
-            np.float16: dtype.Float,
-            np.float32: dtype.Float,
-            np.float64: dtype.Float,
-            tuple: dtype.Tuple,
-            list: dtype.List,
-        }
-        mmap = self.resources.method_map
-        for t1, t2 in type_map.items():
-            for name, prim in mmap[t2].items():
-                method = getattr(t1, name, None)
-                if method is not None:
-                    self.object_map[method] = _Unconverted(prim)
-
-    def __call__(self, value):
-        """Convert a value."""
-        try:
-            v = self.object_map[value]
-            if isinstance(v, _Unconverted):
-                v = self.converter(self, v.value)
-                self.object_map[value] = v
-        except (TypeError, KeyError):
-            v = self.converter(self, value)
-
-        if isinstance(v, Graph):
-            self.resources.manager.add_graph(v)
-        return v
-
-
-class InferenceResource(PipelineResource):
-    """Performs inference and specialization."""
-
-    def __init__(self,
-                 pipeline_init,
-                 tracks,
-                 required_tracks,
-                 tied_tracks,
-                 context_class,
-                 erase_value):
-        """Initialize an InferenceResource."""
-        super().__init__(pipeline_init)
-        self.manager = self.resources.manager
-        self.tracks = tracks
-        self.required_tracks = required_tracks
-        self.tied_tracks = tied_tracks
-        self.context_class = context_class
-        self.erase_value = erase_value
-        self.engine = InferenceEngine(
-            self.pipeline,
-            tracks=self.tracks,
-            tied_tracks=self.tied_tracks,
-            context_class=self.context_class,
-        )
-
-    def fill_in(self, argspec):
-        """Fill in argspec with values for all tracks.
-
-        The 'value' track will also be filled in even if it already exists,
-        since it needs to be wrapped for the inferrer to use it.
-        """
-        for arg in argspec:
-            if '_erase_value' in arg:
-                erase = arg['_erase_value']
-                del arg['_erase_value']
-            else:
-                erase = self.erase_value
-            if 'value' in arg:
-                v = arg['value']
-                for track_name, track in self.engine.tracks.items():
-                    if track_name not in arg:
-                        arg[track_name] = track.from_value(v, None)
-                    else:
-                        arg[track_name] = track.from_external(arg[track_name])
-                if erase:
-                    arg['value'] = ANYTHING
-            else:
-                for track_name, track in self.engine.tracks.items():
-                    if track_name in arg:
-                        arg[track_name] = track.from_external(arg[track_name])
-                    else:
-                        arg[track_name] = track.default(arg)
-
-    def infer(self, graph, argspec, outspec=None, clear=False):
-        """Perform inference."""
-        if clear:
-            self.engine.cache.clear()
-            for node in self.manager.all_nodes:
-                orig_t = node.type
-                node.inferred = defaultdict(lambda: UNKNOWN)
-                if node.is_constant() \
-                        and ismyiatype(orig_t) \
-                        and not ismyiatype(orig_t, Function):
-                    node.type = orig_t
-        self.fill_in(argspec)
-        return self.engine.run(
-            graph,
-            argspec=argspec,
-            outspec=outspec,
-            tracks=self.required_tracks
-        )
-
-    def specialize(self, graph, context):
-        """Perform specialization."""
-        spc = TypeSpecializer(self.engine)
-        result = spc.run(graph, context)
-        self.manager.keep_roots(result)
-        return result
-
-    def renormalize(self, graph, argspec, outspec=None):
-        """Perform inference and specialization."""
-        _, context = self.infer(graph, argspec, outspec, clear=True)
-        return self.specialize(graph, context)
-
-
-class Parser(PipelineStep):
-    """Pipeline step to parse a function.
-
-    Inputs:
-        input: A function.
-
-    Outputs:
-        graph: A graph.
-    """
-
-    def step(self, input):
-        """Assert that input is a Graph, and set it as the 'graph' key."""
-        g = self.resources.convert(input)
-        assert isinstance(g, Graph)
-        return {'graph': g}
-
-
-class Optimizer(PipelineStep):
-    """Pipeline step to optimize a graph.
-
-    Inputs:
-        graph: The graph to optimize.
-
-    Outputs:
-        graph: The optimized graph.
-    """
-
-    def __init__(self,
-                 pipeline_init,
-                 phases,
-                 run_only_once=False):
-        """Initialize an Optimizer."""
-        super().__init__(pipeline_init)
-        self.run_only_once = run_only_once
-        self.phases = []
-        for name, spec in phases.items():
-            if spec == 'renormalize':
-                pass
-            elif isinstance(spec, list):
-                spec = PatternEquilibriumOptimizer(*spec, optimizer=self)
-            else:
-                spec = spec(optimizer=self)
-            self.phases.append(spec)
-
-    def step(self, graph, argspec=None, outspec=None):
-        """Optimize the graph using the given patterns."""
-        changes = True
-        while changes:
-            changes = False
-            for opt in self.phases:
-                if opt == 'renormalize':
-                    assert argspec is not None
-                    graph = self.resources.inferrer.renormalize(
-                        graph, argspec, outspec
-                    )
-                elif opt(graph):
-                    changes = True
-            if self.run_only_once:
-                break
-        self.resources.manager.keep_roots(graph)
-        return {'graph': graph}
-
-
-class InferrerStep(PipelineStep):
-    """Pipeline step to run type/shape/value/etc. inference.
-
-    Inputs:
-        graph: The graph to infer.
-        argspec: Information about argument types.
-
-    Outputs:
-        outspec: Inference results for the graph's output.
-        inferrer: The inference engine.
-    """
-
-    def step(self, graph, argspec):
-        """Infer types, shapes, values, etc. for the graph."""
-        try:
-            res, context = self.resources.inferrer.infer(graph, argspec)
-            return {'outspec': res,
-                    'inference_context': context}
-        except Exception as exc:
-            # We still want to keep the inferrer around even
-            # if an error occurred.
-            return {'error': exc,
-                    'error_step': self}
-
-
-class Specializer(PipelineStep):
-    """Pipeline step to specialize a graph.
-
-    Inputs:
-        graph: The graph to specialize.
-        inferrer: The inference engine.
-
-    Outputs:
-        graph: The specialized graph.
-    """
-
-    def step(self, graph, inference_context):
-        """Specialize the graph according to argument types."""
-        new_graph = self.resources.inferrer.specialize(
-            graph, inference_context
-        )
-        return {'graph': new_graph}
-
-
-class Preparator(PipelineStep):
-    """Pipeline step to prepare the graph to optimization.
-
-    This should be run on the specialized graph.
-
-    Inputs:
-        graph: The graph to prepare.
-
-    Outputs:
-        graph: The prepared graph.
-    """
-
-    def __init__(self,
-                 pipeline_init,
-                 flatten_data_structures=True):
-        """Initialize a Preparator."""
-        super().__init__(pipeline_init)
-        self.flatten_data_structures = flatten_data_structures
-
-    def step(self, graph, argspec, outspec):
-        """Prepare the graph."""
-        orig_argspec = argspec
-        orig_outspec = outspec
-        if self.flatten_data_structures:
-            mng = self.resources.manager
-            erase_class(graph, mng)
-            argspec = tuple(dict(p.inferred) for p in graph.parameters)
-            graph = self.resources.inferrer.renormalize(graph, argspec)
-            erase_tuple(graph, mng)
-            argspec = tuple(dict(p.inferred) for p in graph.parameters)
-            graph = self.resources.inferrer.renormalize(graph, argspec)
-            outspec = dict(graph.output.inferred)
-        return {'graph': graph,
-                'orig_argspec': orig_argspec,
-                'argspec': argspec,
-                'orig_outspec': orig_outspec,
-                'outspec': outspec}
-
-
-class Validator(PipelineStep):
-    """Pipeline step to validate a graph prior to compilation.
-
-    Inputs:
-        graph: The graph to validate.
-
-    Outputs:
-        None.
-    """
-
-    def __init__(self,
-                 pipeline_init,
-                 whitelist=default_whitelist,
-                 validate_type=default_validate_type):
-        """Initialize a Validator."""
-        super().__init__(pipeline_init)
-        self.whitelist = whitelist
-        self.validate_type = validate_type
-
-    def step(self, graph, argspec=None, outspec=None):
-        """Validate the graph."""
-        graph = self.resources.inferrer.renormalize(
-            graph, argspec, outspec
-        )
-        validate(graph,
-                 whitelist=self.whitelist,
-                 validate_type=self.validate_type)
-        return {'graph': graph}
-
-
-class ClosureConverter(PipelineStep):
-    """Pipeline step to closure convert a graph.
-
-    Inputs:
-        graph: The graph to closure convert.
-
-    Outputs:
-        graph: The closure converted graph.
-    """
-
-    def step(self, graph):
-        """Closure convert the graph."""
-        closure_convert(graph)
-        return {'graph': graph}
-
-
-class DebugVMExporter(PipelineStep):
-    """Pipeline step to export a callable.
-
-    Inputs:
-        graph: The graph to wrap into a callable.
-
-    Outputs:
-        output: The callable.
-    """
-
-    def __init__(self, pipeline_init, implementations):
-        """Initialize an DebugVMExporter."""
-        super().__init__(pipeline_init)
-        self.vm = VM(self.pipeline.resources.convert,
-                     self.pipeline.resources.manager,
-                     self.pipeline.resources.py_implementations,
-                     implementations)
-
-    def step(self, graph):
-        """Make a Python callable out of the graph."""
-        return {'output': self.vm.export(graph)}
-
-
-@overload
-def _convert_arg(arg, orig_t: Tuple):
-    if not isinstance(arg, tuple):
-        raise TypeError('Expected tuple')
-    oe = orig_t.elements
-    if len(arg) != len(oe):
-        raise TypeError(f'Expected {len(oe)} elements')
-    return list(flatten(convert_arg(x, o)
-                        for x, o in zip(arg, oe)))
-
-
-@overload  # noqa: F811
-def _convert_arg(arg, orig_t: List):
-    if not isinstance(arg, list):
-        raise TypeError('Expected list')
-    ot = orig_t.element_type
-    return [list(flatten(convert_arg(x, ot) for x in arg))]
-
-
-@overload  # noqa: F811
-def _convert_arg(arg, orig_t: Class):
-    dc = tag_to_dataclass[orig_t.tag]
-    if not isinstance(arg, dc):
-        raise TypeError(f'Expected {dc.__qualname__}')
-    arg = tuple(getattr(arg, attr) for attr in orig_t.attributes)
-    oe = list(orig_t.attributes.values())
-    return list(flatten(convert_arg(x, o)
-                        for x, o in zip(arg, oe)))
-
-
-@overload  # noqa: F811
-def _convert_arg(arg, orig_t: Array):
-    if not isinstance(arg, np.ndarray):
-        raise TypeError('Expected ndarray')
-    et = orig_t.elements
-    assert ismyiatype(et, Number)
-    dtype = type_to_np_dtype(et)
-    if arg.dtype != dtype:
-        raise TypeError('Wrong dtype')
-    return [arg]
-
-
-@overload  # noqa: F811
-def _convert_arg(arg, orig_t: Int):
-    if not isinstance(arg, int):
-        raise TypeError(f'Expected int')
-    return [arg]
-
-
-@overload  # noqa: F811
-def _convert_arg(arg, orig_t: Float):
-    if not isinstance(arg, float):
-        raise TypeError(f'Expected float')
-    return [arg]
-
-
-@overload  # noqa: F811
-def _convert_arg(arg, orig_t: Bool):
-    if not isinstance(arg, bool):
-        raise TypeError(f'Expected bool')
-    return [arg]
-
-
-def convert_arg(arg, orig_t):
-    """Check that arg matches orig_t, and convert to vm_t."""
-    return _convert_arg[orig_t](arg, orig_t)
-
-
-@overload
-def _convert_result(res, orig_t, vm_t: Class):
-    dc = tag_to_dataclass[orig_t.tag]
-    oe = orig_t.attributes.values()
-    ve = vm_t.attributes.values()
-    tup = tuple(convert_result(getattr(res, attr), o, v)
-                for attr, o, v in zip(orig_t.attributes, oe, ve))
-    return dc(*tup)
-
-
-@overload  # noqa: F811
-def _convert_result(res, orig_t, vm_t: List):
-    ot = orig_t.element_type
-    vt = vm_t.element_type
-    return [convert_result(x, ot, vt) for x in res]
-
-
-@overload  # noqa: F811
-def _convert_result(res, orig_t, vm_t: Tuple):
-    # If the EraseClass opt was applied, orig_t may be Class
-    orig_is_class = ismyiatype(orig_t, Class)
-    if orig_is_class:
-        oe = orig_t.attributes.values()
-    else:
-        oe = orig_t.elements
-    ve = vm_t.elements
-    tup = tuple(convert_result(x, o, v)
-                for x, o, v in zip(res, oe, ve))
-    if orig_is_class:
-        dc = tag_to_dataclass[orig_t.tag]
-        return dc(*tup)
-    else:
-        return tup
-
-
-@overload  # noqa: F811
-def _convert_result(arg, orig_t, vm_t: (Int, Float, Bool, Array)):
-    return arg
-
-
-def convert_result(res, orig_t, vm_t):
-    """Convert result from vm_t to orig_t."""
-    return _convert_result[vm_t](res, orig_t, vm_t)
-
-
-class OutputWrapper(PipelineStep):
-    """Pipeline step to convert to and from the VM's data format.
-
-    For example, dataclasses in the arguments may be converted to
-    tuples, and tuples returned by the VM would be converted back
-    to the appropriate dataclasses, as determined by Myia.
-
-    Inputs:
-        graph: The root graph.
-        output: The callable returned by the VM.
-        argspec: Contains the original types/etc. of the arguments.
-        outspec: Contains the original return type.
-
-    Outputs:
-        output: The wrapped callable.
-    """
-
-    def step(self,
-             graph,
-             output,
-             argspec,
-             outspec,
-             orig_argspec=None,
-             orig_outspec=None):
-        """Convert args to vm format, and output from vm format."""
-        prep_step = getattr(self.pipeline.steps, 'prepare', None)
-        if prep_step is None or not prep_step.flatten_data_structures:
-            raise AssertionError(
-                'OutputWrapper step requires prepare.flatten_data_structures'
-            )
-        fn = output
-        orig_arg_t = [arg['type'] for arg in orig_argspec or argspec]
-        orig_out_t = (orig_outspec or outspec)['type']
-        vm_out_t = graph.type.retval
-
-        def wrapped(*args):
-            args = tuple(flatten(convert_arg(arg, ot) for arg, ot in
-                                 zip(args, orig_arg_t)))
-            res = fn(*args)
-            res = convert_result(res, orig_out_t, vm_out_t)
-            return res
-        return {'output': wrapped}
-
-
-step_parse = Parser.partial()
-
-
-step_resolve = Optimizer.partial(
-    run_only_once=True,
-    phases=dict(
-        resolve=[optlib.resolve_globals]
-    )
-)
-
-
-step_infer = InferrerStep.partial()
-
-
-step_specialize = Specializer.partial()
-
-
-step_prepare = Preparator.partial(
-    flatten_data_structures=True,
-)
-
-
-step_debug_opt = Optimizer.partial(
-    phases=dict(
-        main=[
-            # Branch culling
-            optlib.simplify_always_true,
-            optlib.simplify_always_false,
-
-            # Safe inlining
-            optlib.inline_core,
-            optlib.simplify_partial,
-
-            # Miscellaneous
-            optlib.elim_j_jinv,
-            optlib.elim_jinv_j,
-        ],
-        grad=[
-            optlib.expand_J,
-        ],
-        renormalize='renormalize',
-        cse=CSE.partial(report_changes=False),
-        jelim=optlib.JElim.partial(),
-    )
-)
-
-
-step_opt = Optimizer.partial(
-    phases=dict(
-        main=[
-            # Branch culling
-            optlib.simplify_always_true,
-            optlib.simplify_always_false,
-            optlib.simplify_switch1,
-            optlib.simplify_switch2,
-            optlib.simplify_switch_idem,
-            optlib.combine_switches,
-
-            # Safe inlining
-            optlib.inline_trivial,
-            optlib.inline_unique_uses,
-            optlib.inline_inside_marked_caller,
-            optlib.inline_core,
-            optlib.simplify_partial,
-            optlib.replace_applicator,
-
-            # Specialization
-            optlib.specialize_on_graph_arguments,
-
-            # Arithmetic simplifications
-            optlib.multiply_by_one_l,
-            optlib.multiply_by_one_r,
-            optlib.multiply_by_zero_l,
-            optlib.multiply_by_zero_r,
-            optlib.add_zero_l,
-            optlib.add_zero_r,
-
-            # Array simplifications
-            optlib.elim_distribute,
-            optlib.elim_array_reduce,
-
-            # Miscellaneous
-            optlib.elim_identity,
-            optlib.getitem_tuple,
-            optlib.setitem_tuple,
-            optlib.tail_tuple,
-            optlib.elim_j_jinv,
-            optlib.elim_jinv_j,
-            optlib.cancel_env_set_get,
-            optlib.getitem_newenv,
-            optlib.getitem_env_add,
-            optlib.simplify_array_map,
-        ],
-        main2=[
-            # Costlier optimizations
-            optlib.float_tuple_getitem_through_switch,
-            optlib.float_env_getitem_through_switch,
-            optlib.incorporate_getitem,
-            optlib.incorporate_env_getitem,
-            optlib.incorporate_call,
-            optlib.incorporate_getitem_through_switch,
-            optlib.incorporate_env_getitem_through_switch,
-            optlib.incorporate_call_through_switch,
-        ],
-        grad=[
-            optlib.expand_J,
-        ],
-        renormalize='renormalize',
-        cse=CSE.partial(report_changes=False),
-        jelim=optlib.JElim.partial(),
-    )
-)
-
-
-step_validate = Validator.partial()
-
-
-step_cconv = ClosureConverter.partial()
-
-
-step_debug_export = DebugVMExporter.partial(
-    implementations=vm_implementations
-)
-
-
-step_wrap = OutputWrapper.partial()
+from .resources import scalar_object_map, standard_object_map, \
+    standard_method_map, default_convert, ConverterResource, \
+    InferenceResource
+from .steps import step_parse, step_resolve, step_infer, step_specialize, \
+    step_prepare, step_debug_opt, step_opt, \
+    step_validate, step_cconv, step_wrap, step_debug_export
+from .utils import as_frozen
 
 
 standard_resources = dict(
     manager=GraphManager.partial(),
     py_implementations=py_implementations,
     method_map=standard_method_map,
-    convert=Converter.partial(
+    convert=ConverterResource.partial(
         object_map=standard_object_map,
         converter=default_convert
     ),
